backend/pyproject.toml
[tool.poetry]
name = "buho-backend"
version = "0.1.0"
description = "Buho security scan orchestrator backend"
authors = ["Buho Team <dev@buho.local>"]
readme = "../README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
aiofiles = "^23.2.1"
python-dotenv = "^1.0.1"
anyio = "^4.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"
pydantic = "^1.10.14"
pydantic-settings = "^2.2.1"
celery = "^5.3.6"
redis = "^5.0.1"
httpx = "^0.25.2"
structlog = "^24.1.0"
python-multipart = "^0.0.7"
zipfile36 = "^0.1.3"
'psycopg[binary]' = "^3.1.18"
sqlalchemy-utils = "^0.41.1"
alembic = "^1.13.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
